Fix flaky reconnection
Added wait time between attempting a reconnect when a disconnect event
is sent by slack.
