--- close.go
@@ -1,6 +1,0 @@
-package socketmode
-
-// Close cleans up resources
-func (c *Client) Close() error {
-	return c.conn.Close()
-}
