+++ start.go
@@ -23,11 +23,12 @@ func (c *Client) Start(ctx context.Context) error {
 				switch err {
 				case ErrDisconnectRequest:
 					// Retry since this is slack's way of telling the client to retry connection
+					time.Sleep(defaultWaitTimeBetweenRetries)
 					continue
 				}
 
 				// Handle websocket connection error
-				if websocket.IsCloseError(err) && c.retries < c.maxRetries {
+				if websocket.IsUnexpectedCloseError(err) && c.retries < c.maxRetries {
 					c.retries++
 					c.logger.Info("connection retry attempt %d\n", c.retries)
 					time.Sleep(defaultWaitTimeBetweenRetries)
@@ -64,4 +65,28 @@ func (c *Client) start(ctx context.Context) error {
 	c.logger.Info("connection failed")
 
 	return context.Cause(ctx)
+}
+
+// connect grabs the `ws` url using the webapi to make the call and opens up a websocket
+// connection.
+func (c *Client) connect(ctx context.Context) error {
+	c.logger.Info("connecting")
+	_, url, err := c.Api.StartSocketModeContext(ctx)
+	if err != nil {
+		return err
+	}
+
+	if c.debugReconnects {
+		url += "&debug_reconnects=true"
+	}
+
+	c.conn, _, err = websocket.DefaultDialer.Dial(url, nil)
+	if err != nil {
+		return err
+	}
+
+	c.pingTimer = time.NewTimer(c.maxPingInterval)
+	c.conn.SetPingHandler(pingHandlerFunc(c))
+	c.logger.Info("connected")
+	return nil
 }
